added pause menu
added a pause menu to the game, from which you can resume the game or go back to the title screen. There ist still a problem with going back to the title screen from the main game (the cursor doesn't appear and the menu no longer functions correctly). Also, the player still moves around while the game is paused so this has to be corrected as well.